refactor(git-cleanup): thin git-cleanup into phase pipeline (resolves #2496)
