+++ .agents/scripts/lib/orchestration/git-cleanup/phases/branches.js
@@ -1,0 +1,198 @@
+/**
+ * branches.js — branch-reap phase of git-cleanup (Story #2466).
+ * Owns `planCleanup` + `executeCleanup`. Reap helpers live in
+ * `branches-reap.js`.
+ * @module lib/orchestration/git-cleanup/phases/branches
+ */
+
+import {
+  deleteBranchLocal,
+  deleteBranchRemote,
+} from '../../../git-branch-cleanup.js';
+import { Logger } from '../../../Logger.js';
+import {
+  buildPruneSummary,
+  reapLocalRef,
+  reapRemoteRef,
+  reapWorktree,
+} from './branches-reap.js';
+import { computeProtectedReason } from './filters.js';
+import {
+  currentBranch as defaultCurrentBranch,
+  listLocalBranches,
+  listMergedBranches,
+  listRemoteBranches,
+  probeMergedPr,
+  pruneRemoteTracking,
+  readProtectedConfig,
+  removeWorktree,
+  worktreesByBranch,
+} from './git-probes.js';
+import { parsePrunedRefs } from './prune.js';
+
+function evaluateLocalBranch({
+  branch,
+  classify,
+  filter,
+  mergedByGit,
+  prProbe,
+  cwd,
+  wtMap,
+}) {
+  const protectedReason = classify(branch);
+  if (protectedReason) return { skip: { branch, reason: protectedReason } };
+  if (!filter(branch)) return { skip: { branch, reason: 'filtered' } };
+  const prInfo = prProbe(branch, cwd);
+  let detectedBy = null;
+  if (prInfo) detectedBy = 'gh';
+  else if (mergedByGit.has(branch)) detectedBy = 'git-merged';
+  else return { skip: { branch, reason: 'not-merged' } };
+  const wt = wtMap.get(branch);
+  return {
+    candidate: {
+      branch,
+      prNumber: prInfo?.number ?? null,
+      mergedAt: prInfo?.mergedAt ?? null,
+      hasWorktree: !!wt,
+      worktreePath: wt?.path ?? null,
+      detectedBy,
+      localExists: true,
+    },
+  };
+}
+
+function collectRemoteOnlyCandidates({
+  remoteLister,
+  remoteName,
+  cwd,
+  localSet,
+  classify,
+  filter,
+  prProbe,
+}) {
+  const out = [];
+  for (const branch of remoteLister(cwd, remoteName)) {
+    if (localSet.has(branch)) continue;
+    if (classify(branch)) continue;
+    if (!filter(branch)) continue;
+    const prInfo = prProbe(branch, cwd);
+    if (!prInfo) continue;
+    out.push({
+      branch,
+      prNumber: prInfo.number ?? null,
+      mergedAt: prInfo.mergedAt ?? null,
+      hasWorktree: false,
+      worktreePath: null,
+      detectedBy: 'remote-only',
+      localExists: false,
+    });
+  }
+  return out;
+}
+
+/** Pure-ish: enumerate merged-branch candidates. */
+export function planCleanup(ctx) {
+  const {
+    cwd,
+    baseBranch,
+    localLister = listLocalBranches,
+    mergedLister = listMergedBranches,
+    currentBranchFn = defaultCurrentBranch,
+    protectedConfigFn = readProtectedConfig,
+    worktreesFn = worktreesByBranch,
+    prProbe = (b, c) => probeMergedPr(b, c),
+    filter = () => true,
+    includeRemoteOnly = false,
+    remoteLister = listRemoteBranches,
+    remoteName = 'origin',
+  } = ctx;
+  const resolvedCurrent = currentBranchFn(cwd);
+  const resolvedConfigured = protectedConfigFn(cwd);
+  const classify = (branch) =>
+    computeProtectedReason({
+      baseBranch,
+      currentBranch: resolvedCurrent,
+      configured: resolvedConfigured,
+      branch,
+    });
+  const wtMap = worktreesFn(cwd);
+  const mergedByGit = new Set(mergedLister(cwd, baseBranch));
+  const localBranches = localLister(cwd);
+  const localSet = new Set(localBranches);
+  const candidates = [];
+  const skipped = [];
+  for (const branch of localBranches) {
+    const out = evaluateLocalBranch({
+      branch,
+      classify,
+      filter,
+      mergedByGit,
+      prProbe,
+      cwd,
+      wtMap,
+    });
+    if (out.skip) skipped.push(out.skip);
+    else candidates.push(out.candidate);
+  }
+  if (includeRemoteOnly) {
+    candidates.push(
+      ...collectRemoteOnlyCandidates({
+        remoteLister,
+        remoteName,
+        cwd,
+        localSet,
+        classify,
+        filter,
+        prProbe,
+      }),
+    );
+  }
+  return { candidates, skipped };
+}
+
+/** Pure-ish: execute the branch reap plan. */
+export function executeCleanup(ctx) {
+  const {
+    candidates,
+    cwd,
+    remote,
+    removeWorktreeFn = removeWorktree,
+    deleteLocalFn = (b, c) => deleteBranchLocal(b, { cwd: c, force: true }),
+    deleteRemoteFn = (b, c) => deleteBranchRemote(b, { cwd: c }),
+    pruneRemoteFn = (c, r) => pruneRemoteTracking(c, r, parsePrunedRefs),
+    remoteName = 'origin',
+    logger = Logger,
+  } = ctx;
+  const worktrees = [];
+  const local = [];
+  const remoteResults = [];
+  const failures = [];
+  for (const cand of candidates) {
+    if (
+      !reapWorktree({
+        cand,
+        removeWorktreeFn,
+        cwd,
+        logger,
+        worktrees,
+        failures,
+      })
+    )
+      continue;
+    if (!reapLocalRef({ cand, deleteLocalFn, cwd, local, failures })) continue;
+    if (remote)
+      reapRemoteRef({ cand, deleteRemoteFn, cwd, remoteResults, failures });
+  }
+  let prune = null;
+  if (remote && remoteResults.length > 0) {
+    prune = buildPruneSummary({ pruneRemoteFn, cwd, remoteName, failures });
+  }
+  return {
+    worktrees,
+    local,
+    remote: remoteResults,
+    prune,
+    failures,
+    ok: failures.length === 0,
+  };
+}
