+++ .agents/scripts/lib/orchestration/git-cleanup/phases/prompts.js
@@ -1,0 +1,59 @@
+/**
+ * prompts.js — readline-based interactive prompts for git-cleanup
+ * (Story #2466).
+ *
+ * Split out of `cli.js` so each phase file stays under Story #2466's
+ * 200-LOC ceiling. Exports two helpers used by the per-phase drivers:
+ *
+ *   - `promptYesNo(question)` — read a single y/N line and resolve true
+ *     on `y` / `yes` (case-insensitive), false otherwise.
+ *   - `promptStashDecision(entry)` — per-stash drop/keep/quit prompt
+ *     used as the `decideFn` argument to `executeStashes`.
+ *
+ * Both helpers open a fresh readline interface and close it inside a
+ * `try`/`finally` so the process can still exit cleanly.
+ *
+ * @module lib/orchestration/git-cleanup/phases/prompts
+ */
+
+import readline from 'node:readline';
+
+const TAG = '[git-cleanup]';
+
+/* node:coverage ignore next */
+export async function promptYesNo(question) {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+  try {
+    const ans = await new Promise((resolve) => {
+      rl.question(`${question} [y/N] `, (a) => resolve(a));
+    });
+    return /^y(es)?$/i.test(ans.trim());
+  } finally {
+    rl.close();
+  }
+}
+
+/* node:coverage ignore next */
+export async function promptStashDecision(entry) {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+  try {
+    const ans = await new Promise((resolve) => {
+      rl.question(
+        `${TAG} ${entry.ref} (${entry.createdAt}) ${entry.message} — drop/keep/quit [k]? `,
+        (a) => resolve(a),
+      );
+    });
+    const t = (ans ?? '').trim().toLowerCase();
+    if (t === 'd' || t === 'drop' || t === 'y' || t === 'yes') return 'drop';
+    if (t === 'q' || t === 'quit') return 'quit';
+    return 'keep';
+  } finally {
+    rl.close();
+  }
+}
